// src/types/mockData.ts
export type Bill = {
  bill_id: string;
  bill_number: string;
  title: string;
  description: string;
  status: string;
  status_date: string;
  change_hash: string;
  session_id: string;
  status_desc: string;
  committee_id: string;
  committee: string;
  last_action_date: string;
  last_action: string;
  url: string;
  state_link: string | null;
  tsv: unknown;
  full_text: string;
};

export type Person = {
  people_id: string;
  party: string;
  chamber: string;
  state: string;
  name: string;
  first_name: string;
  middle_name: string | null;
  last_name: string;
  party_id: string;
  role_id: string;
  role: string;
  district: string | null;
  title: string;
  phone: string;
  website: string;
  contact_form: string;
  rss_url: string | null;
  twitter_account: string;
  facebook_account: string;
  youtube_account: string | null;
  bluesky_account: string | null;
  twitch_account: string | null;
  instagram_account: string;
  tiktok_account: string | null;
  linkedin_account: string | null;
  discord_account: string | null;
  img: string;
  time_in_office: string;
  district_number: string | null;
  photo_credit: string;
  total_rollcalls: number;
  votes_cast: number;
  missed_votes: number;
  participation_rate: number;
  missed_vote_rate: number;
  primary_bills_sponsored: number;
  bill_passage_rate: number;
  effectiveness_score: number;
  bipartisan_sponsorship_rate: number;
  avg_opposite_party_cosponsors: number;
};